requirements.txt
django==1.11.8 
djangorestframework==3.7.3
psycopg2==2.7.3.2